pyproject.toml
[project]
name = "mongodb-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.14.1",
    "python-dotenv>=1.1.1",
]
